代码工程/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
torch>=1.13.0
pyarrow>=10.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
dash>=2.7.0
jupyter>=1.0.0
pyyaml>=6.0
pathlib>=1.0.0
nltk>=3.8.0

